requirements.txt
google-cloud-vision
pandas